重要更新
功能添加：
1、节点映射和链路映射分别进行
2、实现连续读取多个vnr文件（仅用一个文件测试）
3、实现映射后物理资源属性（cpu、bandwith）的更新

更新日志：
修改embedding.java
把节点和链路映射方法从静态方法改成普通方法，构造函数的参数传入存储虚拟网络文件的文件夹

映射结果文件说明：
节点数 链路数
虚拟节点id 物理节点id 虚拟节点CPU
……
虚拟链路起点 虚拟链路终点 虚拟链路带宽 物理路径
……

下一步计划：
对于到期的vnr归还资源

+++ src/vnr/bandit/Similarity.java
@@ -6,7 +6,7 @@ public class Similarity {
 	 * 作用：计算两个节点之间的相似程度，从cpu和degree的角度评估
 	 * 参数：待比较相似度的两个节点*/
 	public static double cos(Node m,Node n){
-		int mc,nc;
+		double mc,nc;
 		int md,nd;
 		
 		mc=m.getCpu();nc=n.getCpu();
